Change float types to f64

+++ src/poolball.rs
@@ -1,18 +1,18 @@
 use na::{Vector2, Point2};
 
 pub struct Poolball {
-    pub position: Point2<f32>,
-    pub radius: f32,
-    pub mass: f32,
-    pub velocity: Vector2<f32>,
+    pub position: Point2<f64>,
+    pub radius: f64,
+    pub mass: f64,
+    pub velocity: Vector2<f64>,
 }
 
-const MASS: f32 = 0.1;
-const RADIUS: f32 = 1.0;
+const MASS: f64 = 0.1;
+const RADIUS: f64 = 1.0;
 
 impl Poolball {
     // Creates a new Golfball with a initial position and velocity
-    pub fn new(position: Point2<f32>) -> Poolball {
+    pub fn new(position: Point2<f64>) -> Poolball {
         Poolball {
             position: position,
             radius: RADIUS,
@@ -22,7 +22,7 @@ impl Poolball {
     }
 
     // Updates the balls position using its current velocity, then updating velocity
-    pub fn update(&mut self, acceleration: Vector2<f32>, delta_time: f32) {
+    pub fn update(&mut self, acceleration: Vector2<f64>, delta_time: f64) {
         self.position += self.velocity * delta_time;
         self.velocity += acceleration * delta_time;
     }
